Rename and refactor channels.py controller
